[develop] Creando ruta de usuario
