+++ app.js
@@ -1,6 +1,7 @@
 const debug = require('debug')('app:inicio');
 const express = require('express');
 const config = require('config');
+const usuarios = require('./routes/usuarios')
 //Instanciamos express
 const app = express();
 const morgan = require('morgan');
@@ -12,6 +13,9 @@ app.use(express.urlencoded({extended: true}));
 //Middlewares static, me permite acceder mas facilmente a recursos publicos
 app.use(express.static('public'));
 
+app.use('/api/usuarios', usuarios);
+
+
 //configuracion de entorno
 console.log('Aplicación '+ config.get('nombre'));
 
@@ -22,36 +26,6 @@ if(app.get('env') === 'development'){
     debug('Morgan está habilitado');
 };
 
-const usuarios = [
-    {id: 1, nombre:'One'},
-    {id:2, nombre:'Two'},
-    {id:3, nombre:'Three'}
-];
-
-
-app.get('/', (req, res) => {
-    res.send('Hello wordl');
-});
-
-app.get('/api/usuarios', (req, res) => {
-    res.send(['Arriondo', 'Jeremias']);
-});
-
-app.get('/api/usuarios/:id', (req, res) => {
-    let usuario = usuarios.find(user => user.id === parseInt(req.params.id));
-    if (!usuario) {
-        res.status(404).send('El usuario no existe')
-    }
-    res.send(usuario);
-});
-
-app.post('/api/usuarios', (req, res) => {
-    const usuario = {
-        id: usuarios.length + 1,
-        nombre: req.body.nombre
-    };
-    usuarios.push(usuario)
-})
 
 //Creando una variable de entorno
 const port = process.env.PORT || 5000;
